HW-23/scrapper_app/requirements.txt
django==4.1.7
djangorestframework==3.14.0
celery==5.2.7